+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SparkBigQueryUtil.java
@@ -189,9 +189,9 @@ public class SparkBigQueryUtil {
     }
     // need to return timestamp in epoch microseconds
     java.sql.Timestamp timestamp = (java.sql.Timestamp) sparkValue;
-    long epochSecondsAsMicros = (timestamp.getTime() / 1000) * 1_000_000;
-    int micros = timestamp.getNanos() / 1000;
-    return epochSecondsAsMicros + micros;
+    long epochMillis = timestamp.getTime();
+    int micros = (timestamp.getNanos() / 1000) % 1000;
+    return epochMillis * 1000 + micros;
   }
 
   public static int sparkDateToBigQuery(Object sparkValue) {